Rename module _main.py to application.py
